feat: Componentized ChatsMessageText
